% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.mp4.length.R
\name{change.mp4.length}
\alias{change.mp4.length}
\title{Change the length of a .mp4 file}
\usage{
change.mp4.length(time.mult, input, output, overwrite = F)
}
\arguments{
\item{time.mult}{The value by which the time of the movie will be scaled by
.mp4 file. Calculated as current length / desired length

i.e) if the current length is 10 seconds and the desired length is 8 seconds
the percent.diff would be 10/8.}

\item{input}{The .mp4 file to change length of}

\item{output}{The name of the .mp4 file that the copy with the altered length will be saved as.
Cannot use the same name as the origonl version, unless choosing to overwrite original.}

\item{overwrite}{An optional boolean value, if True will delete the original version and replace with the new, altered length version.}
}
\description{
Creates a and saves a copy of a .mp4 file with a different length to that of the original.
The length of the movie is changed by the percent.diff given.
}
\examples{
Changing an mp4 file from 10s to 5s
change.mp4.length(0.5, test.mp4, shorter_test.mp4)

}
