% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear.dev.list.R
\name{clear.dev.list}
\alias{clear.dev.list}
\title{Clears the dev. list}
\usage{
clear.dev.list()
}
\description{
Calls dev.off() x many times, x being the number of items in dev.list()

Used to clear out all plots stored in XQuartz
}
\examples{
clear.dev.list()


}
