% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_weighted_enst.R
\name{density_weighted_enst}
\alias{density_weighted_enst}
\title{Collapse a 3D array into a 2D array and weight by particle density.}
\usage{
density_weighted_enst(data, layer)
}
\arguments{
\item{data}{The list containing the enstrophy or noise values where each entry correlates to a 3D array of values}

\item{layer}{The entry to collapse from the data provided. ie) n=2 is the second entry in the data list provided.}
}
\description{
Return a 'n by n'  2D array from the parsed 'n by n by n' 3D structure.
The data is collapsed along the Z axis and weighted by the population along each column of cells on the Z axis in the 3D array
}
