% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enstrophy_graphs .R
\name{enstrophy_graphs}
\alias{enstrophy_graphs}
\title{Produces graphs from a hdf5 file containing the a halo from SURFS}
\usage{
enstrophy_graphs(
  halo.vec,
  halo.directory,
  graph.directory,
  g.ind = 3,
  n.max = 6,
  snapshot.nums = seq(70, 199)
)
}
\arguments{
\item{halo.vec}{A vector contianing the names of the halo hdf5 files for which graphs will
be produced}

\item{halo.directory}{A string with the location of the directory containing the halos to be retrieved}

\item{graph.directory}{A string containing the path to the directory where the graphs will be saved}

\item{g.ind}{The graph index, a numerical value of 1, 2 or 3 indicating which graph(s) to
be produced.\preformatted{  1: only make graphs of enstrophy against time 
  
  2: only makes graphs of signal to noise ratio
  
  3: make both graphs
}}

\item{n.max}{The maxium layer of the adaptive mesh to be used, sets the Gobal.nmax}

\item{snapshot.nums}{A vector containing all the snapshots from the hdf5 file which are to be
used in making the graphs.}
}
\description{
Generates graphs of a halo(s) for their mean density weighted
enstrophy against look back time of the simulation and or the signal to
noise ratio of the enstrophy.

Multiple halos can be queued and all saved to the same directory.
}
\examples{

make both enstrophy against time and signal to noise for halos 1,2 and 3
halo.vec = c("halo_1", "halo_2", "halo_3")
halo.directory = "/where/halos/are/saved"
graphs.directory = "/where/graphs/will/be/saved"

enstrophy_graphs(halo.vec, halo.directory, graphs.directory, g.ind = 3)


}
