% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Expand a 2D matrix or matrix array by a factor of 'n'}
\usage{
expand(x, n = 3)
}
\arguments{
\item{x}{A matrix or matrix array type object to be expanded}

\item{n}{The scale factor,
which the side lengths of the matrix will be increased by a factor of.}
}
\description{
Scale up 'expand' a matrix "x" by a scale factor of "n"
, return the expanded matrix
i.e.) if x is a 3x3 matrix and n = 3 return a 9x9 matrix where each entry
in x is now equivalent to 4 values within the new 9x9 matrix.
}
