% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.empty.data.R
\name{generate.empty.data}
\alias{generate.empty.data}
\title{Make Empty List of n many 3^(n-1) Arrays}
\usage{
generate.empty.data(max.cells = 3^Global.nmax)
}
\arguments{
\item{max.cells}{The maximum number of cells for one grid length, the largest array in
the list is defined as max.cells^3}
}
\description{
Creates a list of length 'n', where each entry 'n' is an empty
array of dimension = (3^(n-1), 3^(n-1), 3^(n-1)).
Each entry 'n' corresponds to a layer of the adaptive grid, where the level
of refinement increases with n by a factor of 3.

The initial layer is a buffer layer as its just a box with sidelength 3*viral radius
}
