% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.particle.id.table.4.R
\name{generate.particle.id.table.4}
\alias{generate.particle.id.table.4}
\title{Make Table of Particle IDs, Location and Velocities}
\usage{
generate.particle.id.table.4(x, v, species = 2)
}
\arguments{
\item{Halo}{A hdf5 file that has been revrived using SURFsuite already read into
Rstudio using h5read(). Contains all the
particles locations, velocities, species, ID, mass and scale factor.}

\item{Snapshot}{A Boolean value.

TRUE when a halo file has multiple snapshots within it.
If TRUE it will retrieve the halos snapshot_199 (z = 0)
(this is designed for the SURFS L210N1024NR simulation
complied through Surfsuite).

FALSE when halo file does not contain multiple snapshots and this
function will assume that the snapshot is at z=0}

\item{Species}{An intiger value of either 1 or 2.

1 = baryonic particles

2 = dark matter particles}
}
\description{
Returns a data.table of each of the particles of the selected
species in the cluster. The data table contains each particles location and
velocity in SI units relative to the center of Mass
(each particles velocity and location in converted to SI units from
L210N1024NR simulation units).

The particles as assigned an ID by their index in the table.
The table has dimensions = (6, number of particles of selected species)
}
