% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.weighted.enst.R
\name{pop.weighted.enst}
\alias{pop.weighted.enst}
\title{Enstrophy in a 2D array, weighted by density}
\usage{
pop.weighted.enst(data, layer)
}
\arguments{
\item{data}{The list containing the enstrophy or noise values where each entry correlates to a 3D array of values}

\item{layer}{The entry to collapse from the data provided. ie) n=2 is the second entry in the data list provided.}
}
\description{
Return a 'n by n'  2D array from the parsed 'n by n by n' 3D structure.
The data is collapsed along the Z axis and weighted by the population along each column of cells on the Z axis in the 3D array
}
