% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R200.calc.R
\name{R200.calc}
\alias{R200.calc}
\title{Calculates The Virial Radius}
\usage{
R200.calc(species = halo$particles$species, mass.frac = c(1.17e+08, 6.29e+08))
}
\arguments{
\item{species}{A list or vector type object containing numeric values for each particle
in the halo's species}

\item{mass.frac}{A vector containing the mass fraction of the individual species of
particle within the halo in simulation units (solar_mass / h).
Naturally set to the mass fraction from SURFS L210N1024NR.\preformatted{         (Baryons, Dark Matter)
}}
}
\description{
Calculates the viral radius (R200) of a halo
in SURFS simulation units of Mpc/h
}
