% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side.by.side.R
\name{side.by.side}
\alias{side.by.side}
\title{Combine two .mp4 files into a single .mp4 file}
\usage{
side.by.side(left, right, output)
}
\arguments{
\item{left}{The .mp4 file which will be played on the left hand side of the new .mp4 file}

\item{right}{The .mp4 file which will be played on the right hand side of the new .mp4 file}

\item{output}{The name of the new .mp4 file which will be created}
}
\description{
Combine two .mp4 files of the same length and resolution into
a single .mp4 file. The two movies play side by side to one another. The
new .mp4 file is saved in the working directory.
}
\examples{
combining two .mp4 files in the working directory,


side.by.side('test_left.mp4', 'test_right.mp4', 'a_new_movie.mp4')

}
