% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_step_calc.R
\name{time_step_calc}
\alias{time_step_calc}
\title{time step needed in the movie}
\usage{
time_step_calc(movie.length, fps, max.lbt = 12.72)
}
\arguments{
\item{movie.length}{The length of the desired movie}

\item{fps}{The frames per second of the desired movie}

\item{max.lbt}{The max look back time of the movie in Gyrs, if from surfsuite it is most likely to be 12.72 Gyrs}
}
\description{
Calculate the time step needed in a movie to get a movie of the desired length and fps input
}
