% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.R
\name{unwrap}
\alias{unwrap}
\title{Unwarp a halo}
\usage{
unwrap(frame, snap = 1, half.box.size = 105)
}
\arguments{
\item{frame}{A list or vector type object containing the partice positions at a given
time}

\item{snap}{A numeric value for the snapshot number}

\item{half.box.size}{An optional value for the box width / 2 in simulaiton units (Mpc/h).
It is naturally set as 105 as the simulation boxes are 210 Mpc/h}
}
\description{
Halos can sometimes get wrapped around the edges of their box and
as a result not be fully connected.
This function takes a snapshot and unwraps the halo.

It is assumed that the box size is 210 Mpc/h
}
