% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.up.R
\name{clean.up}
\alias{clean.up}
\title{Remove the files from the working directory}
\usage{
clean.up(files)
}
\arguments{
\item{files}{List of files to remove from the working directory}
}
\description{
remove the files from the working directory given as an input
}
\examples{
clean.up(c('test_movie.mp4', 'frame_00006.png', 'how_to.pdf'))

}
