% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_it_all.R
\name{do_it_all}
\alias{do_it_all}
\title{Create a movie showing a halo and its turbulence}
\usage{
do_it_all(
  halo.hdf5.file,
  movie.name,
  time.step = 0.025,
  n.max = 7,
  FPS = 60,
  enst.col = NULL,
  keep.frames = T,
  keep.movies = T,
  draw.R200 = F,
  specify.frame.numb = NULL
)
}
\arguments{
\item{halo.hdf5.file}{The path to the hdf5 file containing the halo}

\item{movie.name}{The name of the mp4 file to be created}

\item{time.step}{The interval of time in Gyrs between frames,
by adjusting this value the length of the mp4 file is altered. A lower
time.step value results in more interpolated and extrapolated steps of the halo.}

\item{n.max}{the maximum layer of the adaptive mesh used to calculate the enstrophy,
naturally it is set to 7}

\item{FPS}{The frames per second of the the mp4 file}

\item{enst.col}{The colour palette used in the enstrophy video produced,
if NULL the cubehelix colour palette is used.}

\item{keep.frames}{If True keep each individual frame used to create the mp4 file movies,
saved in temporary files.
If False the temporary files are removed after creating the mp4 files.}

\item{keep.movies}{If True each of the individual mp4 files created will be kept,
the movie of the halo file will be saved as <movie.name>.SURFS.mp4 and the
enstrophy movie will be saved as <movie.name>.ENST.mp4}

\item{draw.R200}{If True a circle showing the virial radius of the halo will be overlayed on
each of the frames used on the mp4 files produced.
The R200 shown is scaled by the scale factor.}

\item{specify.frame.numb}{An optional vector, giving the frame numbers to be used to generate images
for the movie.}
}
\description{
Using a hdf5 file created by surfsuite of a halo from the SURFS
simulation an mp4 file is created. This mp4 file is a movie of a side by
side comparison of the halos evolution and enstrophy within the halo.
Enstrophy is a proxy value for the halos turbulence (numerical approximation
of solenoidal turbulence).

The video on the left is the halo showing Dark Matter
(Red) and Baryons (Blue).
The video on the right is the enstrophy
}
\examples{
Creating a side by side movie with R200 being shown and a timestep of 0.02 Gyrs.
do.it.all(halo.hdf5.file= "/Users/..../halo4.hdf", movie.name = "halo4_test", time.step=0.02, draw.R200 = T)

A movie where if in full it would be 10s at 60 fps, therefore 600 frames.
However to have the movie only be on the final 5 seconds of this whole movie:
do.it.all(halo.hdf5.file= "/Users/..../halo4.hdf", movie.name = "halo4_test", specify.frame.numb = seq(499, 600))

}
